refactor: rename CW proxy's get-org to get-cxid
Ref. metriport/metriport-internal#2860
